+++ Makefile
@@ -26,7 +26,7 @@ LIBFT_DIR	= includes/libft/
 commit_msg	= "auto commit"
 
 
-SRCS_LIST	= main.c parse_input.c utils.c ft_handle_signals.c env.c export.c
+SRCS_LIST	= main.c parse_input.c utils.c ft_handle_signals.c env.c export.c updates_lists.c
 
 SRCS 		= $(addprefix $(SRCS_DIR), $(SRCS_LIST))
 OBJS 		= $(addprefix $(OBJS_DIR), $(SRCS_LIST:.c=.o))
